Improved no new notifications message on profile view. Removed the notification's array datastructure from profile view.

+++ messaging/ProfileTopViewCell.m
@@ -102,7 +102,7 @@ const CGFloat PROFILE_TOP_VIEW_HEIGHT = 238;
 
 #pragma mark - Modifiers
 
--(void)showNotificationBubbleWithNotificationCount:(int)notificationCount
+-(void)showNotificationBubbleWithNotificationCount:(NSInteger)notificationCount
 {
     [_notificationImageView setHidden:NO];
     [_notificationLabel setHidden:NO];
